separate EventSourcedAggregateSnapshottingTests from EventSourcedAggregateTests
