+++ Domain.Tests/Domain.Tests.csproj
@@ -145,6 +145,7 @@
     <Compile Include="EventHandlerTests.cs" />
     <Compile Include="EventHandlingTests.cs" />
     <Compile Include="EventSequenceTests.cs" />
+    <Compile Include="EventSourcedAggregateSnapshottingTests.cs" />
     <Compile Include="EventSourcedAggregateTests.cs" />
     <Compile Include="EventSourcedRepositoryMigrationTests.cs" />
     <Compile Include="EventSourcedRepositoryTests.cs" />
